Refactoring 9/15
Removed the visual tree re-enter logic.
Introduced a new interface allowing a viewModel to notify the behavior
via an event when the out animation shall play.

Also introduced the DataContextProvider with an IDataContextProvider
interface.
It allows for a way to find the DataContext that implements the
necessary interface inside of a nested tree like structure.

It's also possible to provide a custom implementation.